Renaming entrypoint.
